Move per-test setup and cleanup into Kleene.Test.ConnCase
The same setup block was copy-pasted into six test modules, and had drifted:
volume_test.exs ran the baseline comparison *before* cleanup rather than after,
metadata_test.exs did not capture the cleanup logs, and network_test.exs and
exec_test.exs skipped the baseline comparison altogether.

ConnCase now owns all of it, so every test starts from a clean host and is
checked against the pre-test snapshot afterwards. network_test.exs and
exec_test.exs pass with the comparison enabled -- they were not opted out for a
reason, they just never had it.

'cleanup/0' and 'compare_to_baseline_environment/1' move from 'test_helper.exs'
to 'Kleened.Test.Utils': ConnCase is compiled (test/utils is in elixirc_paths),
so calling TestHelper from it would warn about an undefined module.
